+++ src/modules/ensemblestatevector_mod.f90
@@ -2475,12 +2475,12 @@ CONTAINS
 
           ! do any required interpolation
           if (horizontalInterpNeeded .and. verticalInterpNeeded) then
-            call int_hInterpolate_r4(statevector_file_r4, statevector_hint_r4)
-            call int_vInterpolate_r4(statevector_hint_r4, statevector_member_r4,         &
+            call int_hInterp_gsv_r4(statevector_file_r4, statevector_hint_r4)
+            call int_vInterp_gsv_r4(statevector_hint_r4, statevector_member_r4,         &
                                      Ps_in_hPa_opt=.true.,checkModelTop_opt=checkModelTop)
 
           else if (horizontalInterpNeeded .and. .not. verticalInterpNeeded) then
-            call int_hInterpolate_r4(statevector_file_r4, statevector_member_r4)
+            call int_hInterp_gsv_r4(statevector_file_r4, statevector_member_r4)
 
           else if (.not. horizontalInterpNeeded .and. verticalInterpNeeded) then
             if (horizontalPaddingNeeded) then
@@ -2488,7 +2488,7 @@ CONTAINS
             else
               call gsv_copy(statevector_file_r4, statevector_hint_r4)
             end if
-            call int_vInterpolate_r4(statevector_hint_r4, statevector_member_r4,         &
+            call int_vInterp_gsv_r4(statevector_hint_r4, statevector_member_r4,         &
                                      Ps_in_hPa_opt=.true.,checkModelTop_opt=checkModelTop)
 
           else if (horizontalPaddingNeeded) then